*.go: make provisioner work at basic level
This commit brings the namespace-provisioner to a minimal working state.
It still does not implement everything in the README.md, however, at
least now the API produces valid kubeconfigs and deletes namespaces
after the TTL (as long as the process isn't killed). This has been
tested against a Kubernetes v1.26 cluster.

Next work will be to:
* include RBAC files for deploying the service;
* migrate the service to an controller pattern utilizing reconciliation
  loops;
* move CI to GitHub Actions;
* add e2e tests using Kind.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>