configuration: make namespace-provisioner more secure
In order to prevent privilege escalations, this commit changes the way
roles are granted to created service accounts. Previously, the
namespace-provisioner would create a role in each new namespace and
bind that role to a new service account. This process would require the
namespace-provisioner to have the permissions included in that role on a
cluster-wide level. That means that if the granted role gave the
namespace-provisioner permission to create pods, then the namespace
provisioner itself would need a cluster role permitting it to create
pods in any namespace.

This is:
1. hard to administer, because it means that the namespace-provisioner's
   own cluster role must be modified to include cluster-wide permissions
   for the granted role and
2. very insecure because it gives the namespace provisioner permissions
   to do things it shouldn't have to be able to do.

This commit re-imagines the way that permissions are granted to service
accounts by giving the namespace-provisioner only `bind` permissions on
an already existing ClusterRole. In this new architecture, a cluster
administrator creates a ClusterRole called
"namespace-provisioner-grant", containing the permissions that should be
granted to service accounts created by the namespace-provisioner, and
the namespace provisioner creates a namespaced RoleBinding to bind the
grant to a service account.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>